requirements.txt
scikit-learn
python-docx
pdfplumber
streamlit==1.18.*
protobuf==3.20.*
altair==4.0.0